so much refactoring
